+++ src/element/eflame/mod.rs
@@ -1,0 +1,9 @@
+mod status;
+mod frame;
+mod acmd;
+
+pub fn install() {
+	frame::install();
+	status::install();
+	acmd::install();
+}
