+++ src/element/elight/acmd/throws.rs
@@ -1,0 +1,44 @@
+use smash::app::sv_animcmd::*;
+use smash::phx::Hash40;
+use smash::app::lua_bind::*;
+use smash::lib::lua_const::*;
+use smash::app::utility::get_kind;
+use smash::hash40;
+use smash::lua2cpp::*;
+use smashline::*;
+use smash_script::*;
+use smash::lib::{L2CValue, L2CAgent};
+use std::mem;
+use smash::app::*;
+use smash::phx::Vector3f;
+use crate::util::*;
+use super::*;
+pub fn install() {
+	smashline::install_acmd_scripts!(
+		mythra_dashgrab
+	);
+}
+#[acmd_script(
+    agent = "elight",
+    script =  "game_catchdash",
+    category = ACMD_GAME,
+	low_priority)]
+unsafe fn mythra_dashgrab(fighter: &mut L2CAgentBase) {
+    let lua_state = fighter.lua_state_agent;
+		frame(fighter.lua_state_agent, 11.0);
+		if macros::is_excute(fighter) {
+			GrabModule::set_rebound(fighter.module_accessor, /*CanCatchRebound*/ true);
+		}
+		frame(fighter.lua_state_agent, 12.0);
+		if macros::is_excute(fighter) {
+			macros::CATCH(fighter, /*ID*/ 0, /*Bone*/ Hash40::new("top"), /*Size*/ 3.0, /*X*/ 0.0, /*Y*/ 6.6, /*Z*/ 3.0, /*X2*/ Some(0.0), /*Y2*/ Some(6.6), /*Z2*/ Some(11.5), /*Status*/ *FIGHTER_STATUS_KIND_CAPTURE_PULLED, /*Ground_or_Air*/ *COLLISION_SITUATION_MASK_G);
+			macros::CATCH(fighter, /*ID*/ 1, /*Bone*/ Hash40::new("top"), /*Size*/ 1.5, /*X*/ 0.0, /*Y*/ 6.6, /*Z*/ 1.5, /*X2*/ Some(0.0), /*Y2*/ Some(6.6), /*Z2*/ Some(13.0), /*Status*/ *FIGHTER_STATUS_KIND_CAPTURE_PULLED, /*Ground_or_Air*/ *COLLISION_SITUATION_MASK_A);
+		}
+		macros::game_CaptureCutCommon(fighter);
+		wait(fighter.lua_state_agent, 2.0);
+		if macros::is_excute(fighter) {
+			grab!(fighter, *MA_MSC_CMD_GRAB_CLEAR_ALL);
+			WorkModule::on_flag(fighter.module_accessor, /*Flag*/ *FIGHTER_STATUS_CATCH_FLAG_CATCH_WAIT);
+			GrabModule::set_rebound(fighter.module_accessor, /*CanCatchRebound*/ false);
+		}
+}
